requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT==0.3.2
flask-marshmallow==0.13.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.4.2
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
zope.interface==5.1.0
